refactor: Extract generateReports from webpack-plugin to cli-utils
