+++ packages/cli-utils/package.json
@@ -36,6 +36,8 @@
   "dependencies": {
     "@bundle-stats/html-templates": "^4.1.9",
     "@bundle-stats/utils": "^4.1.9",
+    "@bundle-stats/plugin-webpack-filter": "^4.1.9",
+    "@bundle-stats/plugin-webpack-validate": "^4.1.9",
     "core-js": "^3.21.0",
     "find-cache-dir": "^3.1.0",
     "fs-extra": "^11.0.0"
