+++ packages/cli-utils/src/index.ts
@@ -3,9 +3,6 @@ import * as T from './text';
 export * from './baseline';
 export * from './create-artifacts';
 export * from './constants';
+export * from './reports';
 
 export const TEXT = T;
-
-export function getReportInfo(report: any): any {
-  return report?.insights?.webpack?.assetsSizeTotal?.data;
-}
