--- packages/webpack-plugin/locales.json
@@ -1,3 +1,0 @@
-{
-  "WEBPACK_CONFIGURATION_URL": "https://github.com/relative-ci/bundle-stats/tree/master/packages/webpack-plugin#bundlestatswebpackpluginoptions"
-}
