+++ packages/webpack-plugin/package.json
@@ -47,7 +47,6 @@
   "homepage": "https://github.com/relative-ci/bundle-stats/blob/master/packages/webpack-plugin#readme",
   "dependencies": {
     "@bundle-stats/cli-utils": "^4.1.9",
-    "@bundle-stats/utils": "^4.1.9",
     "core-js": "^3.21.0",
     "lodash": "^4.17.0"
   },
