+++ packages/webpack-plugin/src/config.js
@@ -1,0 +1,2 @@
+export const OPTIONS_URL =
+  'https://github.com/relative-ci/bundle-stats/tree/master/packages/webpack-plugin#bundlestatswebpackpluginoptions';
