+++ packages/webpack-plugin/typings.d.ts
@@ -1,3 +1,5 @@
+import { ReportOptions } from '@bundle-stats/cli-utils';
+
 export class BundleStatsWebpackPlugin {
   constructor(options?: Partial<BundleStatsWebpackPlugin.Options>) {}
 
@@ -37,37 +39,7 @@ declare namespace BundleStatsWebpackPlugin {
     builtAt?: Boolean;
   }
 
-  interface Options {
-    /**
-     * Use local saved stats for comparison
-     * Default: `true`.
-     */
-    compare?: Boolean;
-
-    /**
-     * Save current webpack stats as baseline
-     * Default: `false`.
-     */
-    baseline?: Boolean;
-
-    /**
-     * Output html report
-     * Default: `true`.
-     */
-    html?: Boolean;
-
-    /**
-     * Output json report
-     * Default: `false`.
-     */
-    json?: Boolean;
-
-    /**
-     * Output directory relative to webpack `output.path`
-     * Default: `''`.
-     */
-    outDir?: String;
-
+  interface Options extends ReportOptions {
     /**
      * webpack stats options
      * Default: `{ assets: true: true, chunks: true, modules: true, hash: true, builtAt: true }`.
